Glassdoor Job Stats

Glassdoor Job Stats

gd_job_stats(..., query = NULL, employer = NULL, location = NULL,
  city = NULL, state = NULL, country = NULL, fromAge = NULL,
  jobType = NULL, minRating = NULL, radius = NULL,
  job_title = NULL, job_category = NULL, returnCities = NULL,
  returnStates = NULL, returnJobTitles = NULL,
  returnEmployers = NULL, admLevelRequested = NULL)

Arguments

...
arguments to pass to gd_api

query
Additional options to pass to the query other than those specified here

employer
Scope the search to a specific employer by specifying the name here.

location
Scope the search to a specific location by specifying it here - city, state, or country.

city
Scope the search to a specific city by specifying it here.

state
Scope the search to a specific state by specifying it here.

country
Scope the search to a specific country by specifying it here.

fromAge
Scope the search to jobs that are less than X days old (-1 = show all jobs (default), 1 = 1 day old, 7 = 1 week old, 14 = 2 weeks old, etc.)

jobType
Scope the search to certain job types. Valid values are all (default), fulltime, parttime, internship, contract, internship, temporary

minRating
Scope the search to jobs of companies with rating >= minRating (0 = returns all (default), 1 = more than 1 star, 2 = more than 2 stars, 3 = more than 3 stars, 4 = more than 4 stars)

radius
Scope the search to jobs within a certain radius, in miles, of the location specified.

job_title
Scope the search to a specific job title by specifying it here.

job_category
Job category id to scope the search to - see the Job Category table below - note you must pass the id. This can be a comma separated list of ids if you desire to select more than one category.

returnCities
Results will include geographical data (job counts) broken down by city.

returnStates
Results will include geographical data (job counts, score) broken down by the type of geographical district specified in parameter admLevelRequested.

returnJobTitles
Results will include job data broken down by job title.

returnEmployers
Results will include job data broken down by employer.

admLevelRequested
Geographic district type requested when returnStates is true (1 = states, 2 = counties)

Examples

if (have_gd_tokens()) {

paste0("returnStates=true&admLevelRequested=1")
res = gd_job_stats(returnStates = TRUE, admLevelRequested = 1)
}

#> Warning: Not Found (HTTP 404).