General Glassdoor API function

General Glassdoor API function

glassdoor_api(action = NULL, other = NULL, agent = glassdoor_user_agent(),
  version = 1, q = NULL, format = "json", url = glassdoor_url(),
  pid = glassdoor_pid(), pat = glassdoor_pat(), ip_address = NULL,
  query = NULL, ...)

Arguments

action

The particular API call that you would like to make

other

Each API action will require different parameters For example, an employerId is passed in order to retrieve reviews.

agent

The User-Agent (browser) of the end user to whom the API results will be shown. Note that you can can obtain this from the "User-Agent" HTTP request header from the end-user

version

The API version. The current version is 1 except for jobs, which is currently version 1.1

q

Query parameter (not a list) for Glassdoor

format

Either xml or json as you prefer

url

Glassdoor RESTful API URL

pid

Your partner id, as assigned by Glassdoor

pat

Your partner key, as assigned by Glassdoor

ip_address

The IP address of the end user to whom the API results will be shown

query

Additional options to pass to the query other than those specified here

...

Additional options to send to GET

Value

A list of class glassdoor_api

Examples

if (have_glassdoor_pid() && have_glassdoor_pat()) {
res = glassdoor_api(
action = "employers",
other = NULL,
  version = 1,
format = "json",
query = list(q = "pharmaceuticals"))

res = glassdoor_api(
action = "employers",
other = NULL,
  version = 1,
format = "json",
q = "pharmaceuticals", config = list())
}

#> No encoding supplied: defaulting to UTF-8.

#> No encoding supplied: defaulting to UTF-8.